scheduler_data/scheduler/requirements.txt
pandas
snowflake-connector-python
snowflake-sqlalchemy
